Solution to Calculate and Print the Area of a Circle

+++ Calculate and Print the Area of a Circle@@6788d5012db25efec946f943/code.c
@@ -1,9 +1,9 @@
 #include <stdio.h>
 int main(){
-    float r;
+    int r;
     float a;
     scanf("%f",&r);
     a=3.14*r*r;
-    printf("Area: %f",a);
+    printf("Area: %.2f",a);
     return 0;
 }
